Show an AlertDialog when clicking Get More
Currently, a WebView is shown with links to the Play Store and the Wiki.
On Smartphones, it is very hard to click the correct link due to their
small size, one can click the wrong one very easily.

This commit shows an AlertDialog to make clicking the right link easier
and less error-prone.

Remove Theme Showcase link as well since the app is no longer available
in the Play Store.

Change-Id: I6abd01383ec64983ff3bf40a2d93280428a3f791