themechooser: listen for previews
Add a ContentObserver to the Browse fragmente to listen for preview changes. If the user
is in the app and a theme is installed,  a new card with the theme info is loaded but previews
could not be loaded yet (there is a system background for that task), so just listen to
previews table updates and refresh the cards in that case.

Change-Id: Ib3d46388fdf94fe12884abbdf17c228d784c5f03
JIRA: CML-149
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>